Create ncaaBball.py
* All data taken from ports-reference.com/cbb/

Each key in the ppg_by_ranking dictionary is an assigned ranking, with each value being a range of possible points scored. Each conference is then initialized as a dictionary with the key being the team name and the value being the range of SRS (Simple Rating System) ratings from the 2011-12 season to the 2021-22 season. SRS, as described in the Sports Reference glossary, "is a rating that takes into account average point differential and strength of schedule." For example, Baylor's highest SRS rating in the ten-year timeframe chosen was 1st in the country while their lowest was 41st. Each team in the NCAA is assigned a random ranking within this range, which is used to determine their range of possible points scored. The value ranges in the ppg_by_ranking dictionary increase by .1 both to allow for different ranges for all 358 possible rankings and to reflect that the average points scored difference for two teams one ranking apart is close to 0, which is reinforced by Sports Reference's data.

Each game in the preseason, conference, conference tournament, and NCAA tournament takes two teams, uses their randomly assigned ranking to find their points scored range and generates a points scored number for each team by taking a sample of size 1 from a uniform distribution. This way, the dominance of higher seeds is evident but not absolute; upsets and Cinderella stories can still occur.